--- constant_data.py
@@ -1,8 +1,0 @@
-# константы для валюты обмена USD
-TICKER_USD = 'USD000UTSTOM'
-FIGI_USD = 'BBG0013HGFT4'
-BKK_USD_FAMILY = 'USD50'
-# константы для рассчета конверсии
-RAIF_EX = 2.257
-SWIFT_RAIF = 3.0
-SWIFT_BKKB = 0.21
