+++ logs/bangkok_bank.log
@@ -1,2 +1,10 @@
 2023-03-27 13:20:39,778 DEBUG LAST BANGKOKBANK UPDATE RATE ['27', '03', '2023'] - 12:45     
 2023-03-27 13:20:39,935 DEBUG T-T RATE LAST UPDATE 34.17
+2023-03-31 10:32:50,711 DEBUG LAST BANGKOKBANK UPDATE RATE ['31', '03', '2023'] - 10:10     
+2023-03-31 10:32:50,977 DEBUG T-T RATE LAST UPDATE 33.97
+2023-03-31 10:33:06,873 DEBUG LAST BANGKOKBANK UPDATE RATE ['31', '03', '2023'] - 10:10     
+2023-03-31 10:33:07,087 DEBUG T-T RATE LAST UPDATE 33.97
+2023-03-31 10:57:36,040 DEBUG LAST BANGKOKBANK UPDATE RATE ['31', '03', '2023'] - 10:10     
+2023-03-31 10:57:36,219 DEBUG T-T RATE LAST UPDATE 33.97
+2023-03-31 10:57:47,586 ERROR GETTING LAST RATES UPDATE FAILED : Access Denied
+2023-03-31 11:25:11,718 ERROR GETTING LAST RATES UPDATE FAILED : Access Denied
