+++ logs/tinkoff_logs.log
@@ -13,3 +13,21 @@
 2023-03-27 13:26:11,332 DEBUG CANDLES INFO FOR FIGI BBG0013HGFT4 HAVE BEEN FOUND
 2023-03-27 13:26:11,334 DEBUG RATES HAVE BEEN RECEIVED
 2023-03-27 13:26:11,336 DEBUG GET RATE WAS ACCOMPLISHED. MAX RATE: 76
+2023-03-31 10:36:59,559 DEBUG ALL FIGI'S LIST HAVE BEEN FOUND
+2023-03-31 10:59:26,983 ERROR authentication token is missing or invalid
+2023-03-31 11:01:26,745 ERROR authentication token is missing or invalid
+2023-03-31 11:02:44,962 ERROR authentication token is missing or invalid
+2023-03-31 11:05:51,120 DEBUG ALL FIGI'S LIST HAVE BEEN FOUND
+2023-03-31 11:09:19,789 DEBUG ALL FIGI'S LIST HAVE BEEN FOUND
+2023-03-31 11:09:28,965 ERROR authentication token is missing or invalid
+2023-03-31 11:10:13,961 ERROR authentication token is missing or invalid
+2023-03-31 11:10:49,342 ERROR authentication token is missing or invalid
+2023-03-31 11:11:53,774 ERROR authentication token is missing or invalid
+2023-03-31 11:15:45,940 ERROR authentication token is missing or invalid
+2023-03-31 11:17:52,602 ERROR authentication token is missing or invalid
+2023-03-31 11:27:06,434 DEBUG CANDLES INFO FOR USD HAVE BEEN FOUND
+2023-03-31 11:27:07,347 ERROR authentication token is missing or invalid
+2023-03-31 11:27:29,512 DEBUG CANDLES INFO FOR USD HAVE BEEN FOUND
+2023-03-31 11:27:33,128 DEBUG CANDLES INFO FOR FIGI BBG0013HGFT4 HAVE BEEN FOUND
+2023-03-31 11:27:33,135 DEBUG RATES HAVE BEEN RECEIVED
+2023-03-31 11:27:34,811 DEBUG GET RATE WAS ACCOMPLISHED. MAX RATE: 77
