+++ src/bin/telebot_from_heroku.py
@@ -1,10 +1,10 @@
 import os
 from dotenv import load_dotenv
 
-from telebot_client_base import TelegramBotClient
+from src.listener.telebot_listener import TelegramBotClient
 
 if __name__ == "__main__":
-    load_dotenv(os.path.abspath('.env'))
+    load_dotenv(os.path.abspath('../../.env'))
     TOKEN = os.environ.get('TELEBOT')
     bot = TelegramBotClient(TOKEN)
     bot.run_heroku_server()
