+++ src/clients/const.py
@@ -1,0 +1,3 @@
+# константы для валюты обмена USD
+TICKER_USD = 'USD000UTSTOM'
+FIGI_USD = 'BBG0013HGFT4'
