+++ src/controllers/const.py
@@ -1,0 +1,4 @@
+RAIF_EX = 2.257
+SWIFT_RAIF = 3.0
+SWIFT_BKKB = 0.21
+BKK_USD_FAMILY = 'USD50'
