+++ src/exceptions/base.py
@@ -1,0 +1,2 @@
+class BaseZDVException(Exception):
+    pass
