+++ src/utils/bad_auth_exception.py
@@ -1,0 +1,5 @@
+from src.exceptions.base import BaseZDVException
+
+
+class BadAuthException(BaseZDVException):
+    pass
